requirements.txt
flask
pymongo
gunicorn
osmnx
pandas
spacy
python-dotenv
flask-login
dnspython
certifi